Initial version for CMakeLists.txt to enable compilation without reliance on the Hall D build system (SCons).
1. Navigate to remoll_solid/generators/halld/bggen
2. Make directory
   mkidr build && cd build
3. CMake with reference to code directory
   cmake ../code
4. Compile
   make

For a standard gfortran compile it uses FLAGS "-m32 -ffixed-line-length-none -funroll-all-loops -fno-f2c -O3". This means that a 32-bit compile is forced, and fixed (72 columns) line length is ignored. The remaining compiler flags are common Fortran optimizations.

The CMakeLists.txt file requires that the cernlib command is accessible in the search path.